requirements.txt
pandas 
beautifulSoup4
selenium
requests
slack_sdk
slack 
schedule
openpyxl
python-dotenv


